requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pycodestyle==2.8.0
python-dotenv==0.19.1
requests==2.26.0
soupsieve==2.3
toml==0.10.2
urllib3==1.26.7
